Brought back select by number of vertices and select loose verts/edges: ---------------------------------------------- Split select by number of vertices and select loose verts/edges into seperate functions.
Previously select by number of vertices was
also accessed through two different UI items,
"Select Quads" and "Select Triangles". Now
it is one function with a integer property
for selecting the size of the face you want
selected.

Also added an option to modify the behavior
of the selection. Can now select whether you
want to select faces that have vertices
equal to, less than or greater than the number
of vertices in operator property.